Updated confirmed item
