+++ src/applications/vaos/api/confirmed_va.json
@@ -455,7 +455,7 @@
         "vvsAppointments": [
           {
             "id": "250afc0b-8551-4a79-bbc9-cf5cdd6575c2",
-            "appointmentKind": "ADHOC",
+            "appointmentKind": "MOBILE_GFE",
             "sourceSystem": "SM",
             "dateTime": "2020-11-22T13:35:00Z",
             "duration": 20,
